requirements.txt
amqp==1.4.6
anyjson==0.3.3
astroid==1.4.4
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bcrypt==1.0.2
BeautifulSoup==3.2.1
billiard==3.3.0.20
boto==2.32.1
celery==3.1.18
certifi==2016.2.28
cffi==1.6.0
chardet==2.1.1
chaussette==1.2
circus==0.11.1
circus-web==0.5
clearbit==0.1.3
colorama==0.3.7
coloredlogs==0.4.3
coverage==3.7.1
coveralls==0.4.2
cryptography==1.3.2
cssselect==0.9.1
cssutils==1.0
dateutils==0.6.6
decorator==4.0.9
diff-match-patch==20121119
dj-stripe==0.8.0
Django==1.9.6
django-aggregate-if==0.4
django-allauth==0.25.2
django-annoying==0.7.7
django-appconf==1.0.2
django-atomic-celery==1.1.2
django-atomic-signals==1.1.0
django-autocomplete-light==2.3.1
-e git+git@github.com:neithere/django-autoslug.git@a16fd5e83063e533e709238522fe1321e24349e1#egg=django_autoslug
django-avatar==2.2.1
-e git+https://github.com/concentricsky/django-basic-models.git@v2.2.0#egg=django-basic-models
django-bootstrap3==7.0.0
django-braces==1.8.1
django-bulk-update==1.0.5
# django-cachemodel==2.1.5
-e git+https://github.com/concentricsky/django-cachemodel.git@v2.1.5#egg=django-cachemodel
django-carton==1.2.1
django-celery==3.1.17
django-celery-email==1.0.3
django-chainable-manager==0.5.0
django-cities-light==2.3.1
django-colorfield==0.1.10
django-compat==1.0.10
django-contenttype-tag==1.0.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
# django-emailtemplates==0.11.0
-e git+https://github.com/concentricsky/django-emailtemplates.git@v0.11.0#egg=django-emailtemplates
django-extensions==1.6.7
django-extra-views==0.6.3
django-filter==0.13.0
django-filters==0.2.1
django-grappelli==2.8.1
django-hijack==2.0.7
django-hstore==1.4.2
django-ignoretests2==0.4.1
django-import-export==0.4.5
django-jenkins==0.18.1
-e git+git@github.com:disqus/django-mailviews.git@e3f4adbea04a134e485624077c0194d48f1b646f#egg=django_mailviews
django-mobi==0.1.7
django-model-utils==2.5
django-mptt==0.6.0
django-notifications-hq==1.0
django-ordered-model==1.1.0
django-parsley==0.5
django-pgjson==0.2.3
django-pgsql-interval-field==0.9.4
django-phonenumber-field==0.7.2
django-picklefield==0.3.2
django-pipeline==1.6.8
django-pipeline-browserify==0.4.0
-e git+git@github.com:kevinmarceloph/django-pjaxr.git@faba1b81c5e877efcdde3fa992c7f642e81a9bd0#egg=django_pjaxr
django-polymorphic==0.9.2
django-post-office==2.0.6
django-redis==3.7.0
django-redis-cache==1.6.5
# django-sky-thumbnails==0.2.9
-e git+https://github.com/concentricsky/django-skythumbnails.git@v0.2.9#egg=django-sky-thumbnails
django-storages==1.4
django-su==0.5.1
django-taggit==0.18.3
# fork with django 1.9 compatibility fix
# django-tenant-schemas==1.6.2
-e git+git@github.com:prokaktus/django-tenant-schemas.git@django1.9-fix#egg=django-tenant-schemas
django-tenants==1.1.5
#django-userswitch==0.2.3
-e git+git@github.com:ikraft/django-userswitch.git@f093b065c377df5867e227c330e6db0818212f9a#egg=django-userswitch
django-uuidfield==0.5.0
django-watson==1.2.1
django-widget-tweaks==1.3
django-wysiwyg-redactor==0.3.8.2
djangorestframework==3.3.3
djangorestframework-hstore==1.1
djangosaml2==0.14.4
djorm-ext-core==0.4.3
djorm-ext-expressions==0.5
-e git+git@github.com:kevinmarceloph/djorm-ext-filtered-contenttypes.git@ef386f7bd141e65007c0eee3b8cb1e374f60b24d#egg=djorm_ext_filtered_contenttypes
djorm-ext-pguuid==0.1
-e git+git@github.com:kevinmarceloph/djorm-pgarray.git@ff0fdd54c0728440dd3a6ba36c762af8f613acf2#egg=djorm_pgarray
docopt==0.6.2
drf-nested-routers==0.9.0
enum34==1.1.6
eventlet==0.12.1
Fabric==1.7.0
feedparser==5.2.1
fhurl==0.1.7
future==0.15.2
futures==3.0.5
gevent==0.13.8
greenlet==0.4.0
hellosign-python-sdk==3.8.5
html2text==2015.4.14
html5lib==0.9999999
httplib2==0.8
idna==2.1
iowait==0.2
ipaddress==1.0.16
ipython==4.1.2
ipython-genutils==0.1.0
Jinja2==2.8
jsonfield==1.0.3
kombu==3.0.26
lazy-object-proxy==1.2.1
lxml==3.5.0
Mako==1.0.4
Markdown==2.3.1
MarkupSafe==0.18
meld3==0.6.10
mimeparse==0.1.3
mixpanel-celery==0.8.0
mock==1.0.1
MonthDelta==0.9.1
newrelic==2.32.0.28
nose==1.3.0
oauth2==1.5.211
oauthlib==1.1.1
paramiko==1.10.1
Paste==2.0.3
path.py==8.1.2
pep8==1.5.6
pexpect==3.0
phonenumbers==7.0.4
pickleshare==0.6
Pillow==3.2.0
pinax-stripe==3.1.0
premailer==2.7.0
psutil==4.0.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6
pycryptodomex==3.4.2
PyExecJS==1.1.0
pyflakes==0.8.1
pylint==1.5.4
pynliner==0.5.1
pyOpenSSL==16.0.0
PyReact==0.5.2
pysaml2==4.0.5
python-dateutil==2.5.3
python-docx==0.8.5
python-gflags==2.0
python-linkedin==4.0
python-memcached==1.48
python-modargs==1.7
python-openid==2.2.5
pytz==2016.4
PyYAML==3.11
pyzmq==15.2.0
raven==5.26.0
redis==2.10.5
repoze.who==2.2
requests==2.10.0
requests-oauthlib==0.6.1
setproctitle==1.1.10
shortuuid==0.4.0
simple-salesforce==0.67.0
simplegeneric==0.8.1
simplejson==3.3.0
singledispatch==3.4.0.3
six==1.10.0
smarturls==0.1.2
sorl-thumbnail==12.3
South==1.0.2
sqlparse==0.1.19
stripe==1.33.0
suds==0.4
tablib==0.11.2
tldextract==1.7.3
tomako==0.1.0
TornadIO2==0.0.4
tornado==4.3
traitlets==4.2.1
twilio==5.4.0
unicodecsv==0.9.4
Unidecode==0.4.19
URLObject==2.3.4
vlk-django-jsonfield==1.2.1
waitress==0.8.9
WebOb==1.6.0
Werkzeug==0.9.6
wrapt==1.10.6
ws4py==0.3.4
xmltodict==0.9.2
YahooBoss-Python==0.2.0
zope.interface==4.1.3
ndg-httpsclient
flower==0.9.1
